+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Changelog
 
 Unreleased
 ----------
+- Add --satellite flag to install script
 
 0.1.1 [revision 76f70b6a72c78fcdf143a19ddcc88a0fbd209b3d]
 ---------------------------------------------------------